Started working on the hardware emulation required.
The basics of the hardware are now in place, such as RAM and a HDD.

The virtual address are translated into the required physical address. With the page data being checked against the page table to check if they have already been loaded into memory.

Page faults have been handled, with the data being read from the disk into memory when a fault occurs.