Added the TLB
Added a TLB into the memory system, this would allow for faster access.

By adding the TLB, it was necessary to add in algorithms to decide on which data to replace.

As such a new package memoryAlgorithms now exists. This features an interface for comunicating with the differnet algortihms, as well as three algortihms, FIFO, LRU and Random. These are used by the TLB in deciding what entry to replace.

The memory.Globals file has bee updated it now included the config information for the system (details such as the size of the TLB and the number of frames) and the system takes the information from there when starting up. It also contians some code that is used in several places for example selecting the replacementAlgorithm.

Exceptions are now passed back through the interface when the occur this will allow for graceful error reporting to the user.

The pageTable has been updated to allow for variable sizes, this means that a page replacment algorithm is also utilised here.

Updates documentation throughout the project.

Created a program to create new input files. This can generate input files that have differing amounts of spatial locality, this means the TLB and PageTables miss rates are not basically 100% as with the sample input.