Statistic tracking
The memory system now provides a way for the callee to view details about the memory system. It provided access to information such as the number of pagefaults incurred.

It also is now fully configurable when it is created so that it is possible to modifiy the number of frames, and the size of the TLB.