Added a GUI to the application.
There are way to many changes to dicuss to make any meaningful commit message.

The core of the project (src/memory) has not chnaged significantly, these means that for the puprose of the assignment nothing much needs said.

The GuI was created with javaFX. Why? I wanted to learn it so thats why. The command line version of the project still exists. (src/textcontrol)